FogBugz #341071: Remove commit 973e60ecef962 to fix SPI1 interrupt
commit 973e60ecef96 ("arm: socfpga: Enable spi and qspi driver for
socfpga") has been around during the initial bringup of C5/A5 devices. The
SPI interrupts were wrong for this commit and have been corrected by the
upstream community. But this commit is getting brought along during
an upgrade cycle, which basically undo's the community's fix.

For v4.4, the commit 973e60ecef96 ("arm: socfpga: Enable spi and qspi
driver for socfpga") should be discarded, but for v4.3, we'll just do
a revert and fix.

This reverts commit 973e60ecef96 ("arm: socfpga: Enable spi and qspi
driver for socfpga").

Signed-off-by: Dinh Nguyen <dinguyen@opensource.altera.com>