stmmac: rework DMA bus setting and introduce new platform AXI structure
[backport 'commit afea03656add ("stmmac: rework DMA bus setting and
introduce new platform AXI structure")']

This patch restructures the DMA bus settings and this is done
by introducing a new platform structure used for programming
the AXI Bus Mode Register inside the DMA module.
This structure can be populated from device-tree as documented in the
binding txt file.

After initializing the DMA, the AXI register can be optionally tuned
for platform drivers based.
This patch also reworks some parameters to make coherent the DMA
configuration now that AXI register is introduced.
For example, the burst_len is managed by using the mentioned axi
support above; so the snps,burst-len parameter has been removed.
It makes sense to provide the AAL parameter from DT to Address-Aligned
Beats inside the Register0 and review the PBL settings when initialize
the engine.

For PCI glue, rebuilding the story of this setting, it
was added to align a configuration so not for fixing some
known problem. No issue raised after this patch.
It is safe to use the default burst length instead of
tuning it to the maximum value

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Signed-off-by: Alexandre TORGUE <alexandre.torgue@st.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Thor Thayer <thor.thayer@linux.intel.com>

---
Thor changes:
	Remove mdio in original patch portion of stmmac.txt
	Replace return plat with return 0 in stmmac_platform.c

Conflicts:
	Documentation/devicetree/bindings/net/stmmac.txt
	drivers/net/ethernet/stmicro/stmmac/stmmac_platform.c